+++ src/app/tf/[species]/[factor]/histone/page.tsx
@@ -3,6 +3,11 @@ import { query } from "../../../../../../lib/client";
 import { DATASETS_QUERY, EPIGENETIC_PROFILE_ACCESSIONS } from "../_ExperimentSelectionPanel/queries";
 import { excludeTargetTypes, includeTargetTypes } from "@/consts";
 
+/**
+ * This file is here to redirect requests to /tf/[species]/[factor]/histone
+ * by adding first experiment to url
+ */
+
 async function getExperiments(processed_assembly: "GRCh38" | "mm10", target: string) {
   return await query({
     query: DATASETS_QUERY,
@@ -25,10 +30,6 @@ async function getValidExps(assembly: "GRCh38" | "mm10") {
   })
 }
 
-/**
- * This file is here to redirect requests to /tf/[species]/[factor]/histone
- * by adding first experiment to url
- */
 export default async function Page({
   params: { species, factor },
 }: {
